# openapi.yaml
openapi: 3.0.3
info:
  title: HabitOS API
  version: 1.0.0
  description: API for HabitOS - Behavioral Optimization Platform
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - timestamp

    # Auth Schemas
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        avatar:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt
        - updatedAt
    UserCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterPayload:
      allOf:
        - $ref: '#/components/schemas/UserCredentials'
        - type: object
          properties:
            name:
              type: string
          required:
            - name
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user:
          $ref: '#/components/schemas/User'
      required:
        - access_token
        - token_type
        - user
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token

    # Behavior Schemas
...
    DashboardStats:
      type: object
      properties:
        totalBehaviors:
          type: integer
        activeBehaviors:
          type: integer
        totalOptimizationRuns:
          type: integer
        completionRate:
          type: number
        averageScore:
          type: number
        streakDays:
          type: integer
      required:
        - totalBehaviors
        - activeBehaviors
        - totalOptimizationRuns
        - completionRate
        - averageScore
        - streakDays
    DashboardSummary:
      type: object
      properties:
        stats:
          $ref: '#/components/schemas/DashboardStats'
        recentOptimizations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              status:
                type: string
              score:
                type: number
              createdAt:
                type: string
                format: date-time
        recentBehaviors:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              category:
                type: string
              isActive:
                type: boolean
              createdAt:
                type: string
                format: date-time
        todaySchedule:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              behaviorName:
                type: string
              timeSlot:
                type: string
              startTime:
                type: string
              isCompleted:
                type: boolean
      required:
        - stats
        - recentOptimizations
        - recentBehaviors
        - todaySchedule
    AnalyticsData:
      type: object
      properties:
        period:
          type: string
        behaviorCompletions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              completed:
                type: integer
              scheduled:
                type: integer
        objectiveProgress:
          type: array
          items:
            type: object
            properties:
              objectiveName:
                type: string
              progress:
                type: number
              trend:
                type: string
                enum: [up, down, stable]
        categoryDistribution:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
              percentage:
                type: number
        energyUsage:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              energySpent:
                type: number
              energyBudget:
                type: number
      required:
        - period
        - behaviorCompletions
        - objectiveProgress
        - categoryDistribution
        - energyUsage
paths:
...
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPayload'
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
  /auth/logout:
    post:
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string

  # Behaviors
...
  /analytics/stats:
    get:
      summary: Get dashboard stats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stats
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardStats'
  /analytics/summary:
    get:
      summary: Get dashboard summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DashboardSummary'
  /analytics:
    get:
      summary: Get detailed analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: period
          schema:
            type: string
            default: "7d"
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AnalyticsData'